renamed Header to HomePage
